Linked List insertion, traversal and deletion in C
Linked list insertion , deletion and traversal coded in C language.
1) Insertion at beginning, insertion at end, insertion at an index
   and insertion after a node.
2) Deletion at beginning, deletion at end, deletion at an index
   and  delete a node.
3) Traversal from head node to end.
