--- C-Programs/Linked-List-C-updated/linked_list.c
@@ -1,24 +1,0 @@
-#include<ll_traversal.h>
-
-int main(){
-    struct Node *head;
-    struct Node *second;
-    struct Node *third;
-
-    head = (struct Node *) malloc(sizeof(struct Node));
-    second = (struct Node *) malloc(sizeof(struct Node));
-    third = (struct Node *) malloc(sizeof(struct Node));
-
-    head->data = 7;
-    head->next = second;
-
-    second->data = 11;
-    second->next = third;
-
-    third->data = 41;
-    third->next = NULL;
-
-    traversal(head);
-
-    return 0; 
-}
