--- C-Programs/Linked-List-C-updated/ll_traversal.h
@@ -1,15 +1,0 @@
-#include<stdio.h>
-#include<stdlib.h>
-
-
-struct Node{
-    int data;
-    struct Node *next;
-};
-
-void traversal(struct Node *ptr){
-    while(ptr != NULL){
-        printf("Element : %d\n",ptr->data);
-        ptr = ptr->next;
-    }
-}
